+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/contentful/contentful.php/compare/1.1.0...HEAD)
 
+### Added
+* Implemented `ResourceArray::getItems` to allow access to the values of a `ResourceArray` as an actual PHP array.
+
 ## [1.1.0](https://github.com/contentful/contentful.php/tree/1.1.0) (2017-05-11)
 
 ### Added